New SQL Statement Management
Completely rewrote how SQL statement parsing works to support class bassed representations of SQL statements with the new `experimental` flag. The classes returned when the experimental flag is on should be completely backwards compatible. Wrote the new system in a way that should also be backwards compatible with the older code in the repository.